@@ -7,8 +7,8 @@
     "axios": "^0.21.1",
     "express": "^4.17.1",
     "nodemon": "^2.0.4",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^7.2.2",
     "react-scripts": "^5.0.1",
     "redux": "^4.0.5",
